Clean up project/dataset/job API tests
